Update references to moved client classes
